requirements.txt
requests==2.26.0
numpy==1.19.2
torchvision==0.8.0
tqdm==4.56.0
scipy==1.6.2
torch==1.7.0
matplotlib==3.3.4
albumentations==1.0.3
cxxfilt==0.2.2
docutils==0.17.1
Pillow==8.3.1
scikit_learn==0.24.2
skimage==0.0
Sphinx==4.1.1
sphinx_rtd_theme==0.5.2
submitit==1.3.3
